Jump test more specific and flexible. Notes on past provided
Made object to hold expected failures and an expected failures
test - Too many things to write specific tests for otherwise. In
the object, I indicated which tests I'd checked over to verify
that they /shouldn't/ be getting the usual input that the single
tests do. Some of them may not be verified for a while. Also
leave notes in there as to why certain things are as they are.
Debating whether to move all expected anomolies into that file.